LayoutTests/platform/chromium-win/plugins/window-geometry-initialized-before-set-window-expected.txt

CONSOLE MESSAGE: PLUGIN: Plugin's HWND has been sized before NPP_SetWindow was called

This tests that a plugin's HWND is already sized and positioned before NPP_SetWindow is called.
